+++ AimSharp/AimSharp.csproj
@@ -8,7 +8,7 @@
     <OutputType>Exe</OutputType>
     <AppDesignerFolder>Properties</AppDesignerFolder>
     <RootNamespace>AimSharp</RootNamespace>
-    <AssemblyName>AimSharp</AssemblyName>
+    <AssemblyName>ShineSharp</AssemblyName>
     <TargetFrameworkVersion>v4.5</TargetFrameworkVersion>
     <FileAlignment>512</FileAlignment>
     <PreBuildEvent />
@@ -71,6 +71,9 @@
     <Reference Include="SharpDX.Direct3D9">
       <HintPath>D:\X\System\SharpDX.Direct3D9.dll</HintPath>
     </Reference>
+    <Reference Include="SPrediction">
+      <HintPath>D:\X\System\SPrediction.dll</HintPath>
+    </Reference>
     <Reference Include="System" />
     <Reference Include="System.Core" />
     <Reference Include="System.Drawing" />
@@ -81,19 +84,11 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="ShineCommon\Activator\Ignite.cs" />
-    <Compile Include="ShineCommon\Activator\Smite.cs" />
-    <Compile Include="ShineCommon\Activator\Summoner.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="ShineCommon\Pool\ObjectPool.cs" />
-    <Compile Include="ShineCommon\Utility.cs" />
-    <Compile Include="ShineCommon\SPrediction\ClipperWrapper.cs" />
-    <Compile Include="ShineCommon\SPrediction\Collision.cs" />
-    <Compile Include="ShineCommon\SPrediction\Geometry.cs" />
-    <Compile Include="ShineCommon\SPrediction\Prediction.cs" />
     <Compile Include="SpellData.cs" />
     <Compile Include="SpellDatabase.cs" />
+    <Compile Include="Utility.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="App.config" />
