--- AimSharp/ShineCommon/Pool/ObjectPool.cs
@@ -1,34 +1,0 @@
-﻿using System;
-using System.Collections.Concurrent;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace ShineCommon
-{
-    public class ObjectPool<T>
-    {
-        private ConcurrentBag<T> _objects;
-        private Func<T> _objectGenerator;
-
-        public ObjectPool(Func<T> objectGenerator)
-        {
-            if (objectGenerator == null) throw new ArgumentNullException("objectGenerator");
-            _objects = new ConcurrentBag<T>();
-            _objectGenerator = objectGenerator;
-        }
-
-        public T GetObject()
-        {
-            T item;
-            if (_objects.TryTake(out item)) return item;
-            return _objectGenerator();
-        }
-
-        public void PutObject(T item)
-        {
-            _objects.Add(item);
-        }
-    }
-
-}
