+++ AimSharp/Utility.cs
@@ -1,0 +1,34 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using LeagueSharp;
+using LeagueSharp.Common;
+
+namespace AimSharp
+{
+    public static class Utility
+    {
+
+        public static string[] HitchanceNameArray = { "Low", "Medium", "High", "Very High", "Only Immobile" };
+        public static HitChance[] HitchanceArray = { HitChance.Low, HitChance.Medium, HitChance.High, HitChance.VeryHigh, HitChance.Immobile };
+
+        
+
+        public static bool IsImmobilizeBuff(BuffType type)
+        {
+            return type == BuffType.Snare || type == BuffType.Stun || type == BuffType.Charm || type == BuffType.Knockup || type == BuffType.Suppression;
+        }
+
+        public static bool IsImmobileTarget(Obj_AI_Hero target)
+        {
+            return target.Buffs.Count(p => IsImmobilizeBuff(p.Type)) > 0 || target.IsChannelingImportantSpell();
+        }
+
+        public static bool IsActive(this Spell s)
+        {
+            return ObjectManager.Player.Spellbook.GetSpell(s.Slot).ToggleState == 2;
+        }
+    }
+}
